Make first-party crates clippy-clean; add clippy CI gate
Fix all clippy warnings across the engine's own crates (mostly mechanical:
collapsible if-let, iterator/enumerate over index loops, is_some_and over
map_or, std::f32::consts for TAU/PI, redundant closures, etc.). Net -300
lines. Judgment calls kept conservative: #[allow(deprecated)] on cpal/egui
APIs whose replacements differ semantically (commented for later migration),
#[allow(dead_code)] on intentional WIP/format scaffolding.

Add a `clippy` job to the Test workflow that gates on `-D warnings` over the
same first-party scope, allowing `too_many_arguments` / `type_complexity`
(inherent to Bevy systems and queries).

+++ crates/renzora_engine/src/debug_log.rs
@@ -87,7 +87,7 @@ pub fn debug_log_cameras(
                         .map(|n| n.to_string())
                         .unwrap_or_else(|| "unnamed".into()),
                     active: cam.is_active,
-                    order: cam.order as isize,
+                    order: cam.order,
                     has_cam3d: cam3d.is_some(),
                     role,
                     has_default: default_cam.is_some(),
